# m3: Reduce allocations when emitting metrics