Similar to #116, emitting metrics requires allocations as TUDPTransport
does not implement TRichTransport, so we end up allocating to convert
`string` to `[]byte`.

Add a new benchmark BenchmarkEmitMetrics to measure the impact. This
change brings down allocations significantly:
```
benchmark                old ns/op     new ns/op     delta
BenchmarkEmitMetrics     119112        90897         -23.69%
benchmark                old allocs     new allocs     delta
BenchmarkEmitMetrics     1297           331            -74.48%
benchmark                old bytes     new bytes     delta
BenchmarkEmitMetrics     5284          3293          -37.68%
```